Move Publish from reports package to udash
Now that I created the udash package, it's just makes more sense to
move the publish method to the udash package.
